refactor: 불필요 엔티티 삭제
