+++ src/main/java/in/koreatech/koin/domain/community/lostitem/chatmessage/model/ChatMessageEntity.java
@@ -2,17 +2,28 @@ package in.koreatech.koin.domain.community.lostitem.chatmessage.model;
 
 import java.time.LocalDateTime;
 
+import org.springframework.data.annotation.Id;
+import org.springframework.data.redis.core.RedisHash;
+import org.springframework.data.redis.core.index.Indexed;
+
+import in.koreatech.koin.domain.community.lostitem.chatmessage.dto.ChatMessage;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
+@AllArgsConstructor
+@RedisHash(value = "ChatMessage")
 public class ChatMessageEntity {
 
+    @Id
     private Long id;
 
+    @Indexed
     private Integer articleId;
 
+    @Indexed
     private Integer chatRoomId;
 
     private Integer userId;
@@ -21,20 +32,23 @@ public class ChatMessageEntity {
 
     private Boolean isDeleted;
 
+    private Boolean isImage;
+
     private String nickName;
 
     private String contents;
 
-    private Boolean isImage;
-
     private LocalDateTime createdAt;
 
-    public void addMessageId(Long id) {
-        this.id = id;
+    public void readMessage() {
+        this.isRead = true;
     }
 
-    public static ChatMessageEntity create(Integer articleId, Integer chatRoomId, Integer userId, Boolean isRead, ChatMessageCommand message) {
+    public static ChatMessageEntity create(
+        Integer articleId, Integer chatRoomId, Integer userId, Boolean isRead, ChatMessage message, Long id
+    ) {
         return ChatMessageEntity.builder()
+            .id(id)
             .articleId(articleId)
             .chatRoomId(chatRoomId)
             .userId(userId)
