+++ src/main/java/in/koreatech/koin/domain/community/lostitem/chatmessage/service/ChatMessageSendService.java
@@ -63,7 +63,7 @@ public class ChatMessageSendService {
             event.articleId(),
             event.chatRoomId(),
             sender.getNickname(),
-            event.messageCommand().content(),
+            event.chatMessage().content(),
             partner
         );
