+++ src/main/java/in/koreatech/koin/domain/community/lostitem/chatroom/service/usecase/LostItemChatRoomInfoUseCase.java
@@ -13,11 +13,11 @@ import in.koreatech.koin._common.code.ApiResponseCode;
 import in.koreatech.koin._common.exception.CustomException;
 import in.koreatech.koin.domain.community.article.model.LostItemArticle;
 import in.koreatech.koin.domain.community.article.repository.LostItemArticleRepository;
+import in.koreatech.koin.domain.community.lostitem.chatmessage.model.ChatMessageEntity;
 import in.koreatech.koin.domain.community.lostitem.chatroom.repository.LostItemChatRoomInfoRepository;
 import in.koreatech.koin.domain.community.lostitem.chatroom.repository.LostItemChatUserBlockRepository;
 import in.koreatech.koin.domain.community.lostitem.chatroom.dto.ChatRoomListResponse;
 import in.koreatech.koin.domain.community.lostitem.chatroom.model.LostItemChatRoomInfoEntity;
-import in.koreatech.koin.domain.community.lostitem.chatmessage.model.ChatMessageEntity;
 import in.koreatech.koin.domain.community.lostitem.chatmessage.repository.ChatMessageRedisRepository;
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.domain.user.repository.UserRepository;
